+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Add Predicate.dsl for building complex expressions without having
   to prefix factory methods with `Predicate.` every time.
 
+* Add Predicate.in(:x, Range) as a shortcut over comparison
+  operators.
+
 # 2.4.0 / 2020-07-23
 
 * Add Predicate#to_hash that allows getting back a Hash object